Add testing and page configurability
This commit introduces a breaking change to the API. The Lambda function is extended to support executing setup code beforehand and assertion tests afterwards. The static webpage has been updated to support the new setup and testing components, and is able to load pre-baked configurations from JSON files based on URL parameter.