Implement configurable exercises and tests (#3)
* Simplify keyboard shortcut

* Make Nu Html Checker happy

* Simplify page layout

* Add testing and page configurability

This commit introduces a breaking change to the API. The Lambda function is extended to support executing setup code beforehand and assertion tests afterwards. The static webpage has been updated to support the new setup and testing components, and is able to load pre-baked configurations from JSON files based on URL parameter.

* Add Lambda tests

The Lambda function can be run directly to test its behaviour with a basic request. The same request is available as a JSON file that can be configured as the test event in AWS Lambda.

* Add interface descriptions to README

* Add page configurations for sample exercises