Make scattered improvements (#5)
* Merge duplicate CSS rule-sets

* Hack to remove quote escaping in variables

* Amend page configurations

* Fix script syntax

* Add random integer substitution

First cut of page configuration preprocessing. It supports replacing placeholders with random integers in [1, 100].

* Rework Exercise 2 configuration

* Show blank setup field on secret switch

* Implement test status indicator in CSS

* Self-host jQuery

* Fix tests passing on errors

* Add read-only setup editor

The setup code has been moved to a read-only editor element with nice scrolling and syntax highlighting. This introduces a regression for live editing in secret mode; as the editor is locked to read-only, it cannot be updated.