Add navigation and negative print tests (#6)
* Fix formatting nitpicks

* Add negative print tests

Breaking API change to add the ability to test that some string was not printed. Exercise 2 page configurations have been updated to use print tests. Note that the print history is still searched as a single string, so substrings of a given line are permitted.

* Add navigation controls

* Add root link to header

* Add navigation config to README

* Remove quotes from non-string variables

Breaking API change to pass additional information required to determine if a stringified variable was originally a string or some other type. Other types will have surrounding quotes removed from view.